updated for tox
Start running tests for multiple versions of Python with tox/detox.

Found odd problem with out-of-order arrays that only surface in
the tox virtualenv. I do not think it relevant to the API wrapper
so I fixed the tests.

When tox creates virtualenvs from requirements-dev.txt, it does
not follow dependencies. Explicitly added what is required for Sphinx.